libziparchive: fix missing fixed-width integer types in zip_writer.h
Compilation fails due to undefined types 'uint16_t' and 'uint32_t' in
zip_writer.h. These types are used in struct members but are not recognized
because the required header <cstdint> is not included. This results in
errors such as:
  error: 'uint16_t' does not name a type
  error: 'uint32_t' does not name a type

Added '#include <cstdint>' to zip_writer.h to ensure fixed-width integer
types are properly defined.

Signed-off-by: Sasi Kumar Maddineni <quic_sasikuma@quicinc.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>