python3-nanobind: Fix build
Fix the order of inherited classes, so the nanobuild python module
actually gets built.

Also, put the cmake files into the correct location, i.e.
  ${libdir}/cmake/nanobind
instead of
  ${base_libdir}/cmake/python3-nanobind

onnx 1.20.0rc1 uses nanobind and during its build it detects the
presence of the python module, and wants to use the nanobind cmake
module, both of which fail.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>