babeld: fix installation with usrmerge
In case usrmerge DISTRO_FEATURE is enabled, the recipe installed
the application to /bin folder, which is however a symlink to /usr/bin,
so the installation ultimately failed.

To fix this, set the correct prefix for the installation.

This is a partial backport of f91983f1f3e897d25ab477fdfb5baed89ccc0daf

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>