cifs-utils: fix broken PACKAGECONFIGs
Some PACKAGECONFIGs (cifsidmap, cifsacl, pam) were failing to build since
a while, erroring out with:

| ../sources/cifs-utils-7.4/resolve_host.c:23:10: fatal error: config.h: No such file or directory
|    23 | #include "config.h"
|       |          ^~~~~~~~~~
| compilation terminated.

The config.h header is generated in the root of build folder, and it seems
that the recipe can't be built 100% out of the source tree.

To avoid this issue, add ${B} as an include folder to CFLAGS, so it finds
the required header.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>