physfs: upgrade 3.0.2 -> latest revision
License-Update: small formatting changes and copyright year update. License hasn't changed.

The previous version of the recipe, 3.0.2 was released in 2019. The last release, 3.2.0 was
done in 2022.
However the project is still actively developed (it just moved to Github) - fetch the latest
revision, at least until there is a new release.

Shortlog:
https://github.com/icculus/physfs/commits/main?since=2019-03-19&until=2025-12-26

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>