tigervnc: compare vendored xserver version to oe-core's xserver
TigerVNC compiles its own xserver from a separate tarball, it doesn't
use oe-core's xserver for most of the compilation.

The vendored xserver code should be still kept in sync with oe-core
to minimize incompatibilities, feature and bug-discrepepancies (...)

However it is easy to miss when xserver is updated in oe-core, TigerVNC's
xserver gets out of sync frequently.

This change adds a small check before do_configure that compares the
vendored and oe-core's xserver version, and issues a warning if
TigerVNC recipe needs to be synced to oe-core.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>